functions/pyproject.toml
[tool.poetry]
name = "functions"
version = "0.1.0"
description = ""
authors = ["Samuel Mallet  <69467005+SuperMuel@users.noreply.github.com>"]
readme = "README.md"
package-mode=false
packages = [
    { include = "src" }
]



[tool.poetry.dependencies]
python = "^3.11"
arrow = "^1.3.0"
validators = "^0.34.0"
pydantic = "^2.9.2"
ics = "^0.7.2"
pytz = "^2024.2"
requests = "^2.32.3"
google-auth-oauthlib = "^1.2.1"
firebase-functions = "^0.2.0"
functions-framework = "^3.8.1"
google-api-python-client = "^2.120.0"
google-auth-httplib2 = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
